refactor(workspace): split create.rs into DDD modules
Break apart the workspace create.rs file (651 lines) into domain-driven
design modules for better separation of concerns:

- editor.rs: EditorType enum and editor launching logic
- path.rs: Path generation functions (sanitize_project_name,
  generate_workspace_path, sanitize_workspace_name,
  generate_standalone_workspace_path, calculate_expires_at) with tests
- data.rs: Data copying functions (copy_issue_data_to_workspace,
  copy_project_config_to_workspace)
- create.rs: Simplified orchestration layer with public API

This follows the existing DDD patterns used in the item/ module.
